Only generate a function definition, if it is used or otherwise visible.
This is a cheap check to avoid some unnecessary work, when headers contain unused static inline functions.
On OS X this prevents creating a function in a system header file, which contains an asm statement with the unsupported constraint "+A".
Using '-O0' this function does not get removed, so 64-bit lowering fails.
This works around the issue by not building IR for the function in the first place.