Allow statement expressions ending in a label containing an expression statement to have this expression as result.
Technically ({ bla: 0; }) does not end in an expression statement, but in a label statement.
Due to an implementation detail in GCC (labels do not have an inner statement, but their inner statement is made their successor) the statement expression does end in an expression statement.

This fixes C/gnu99/stmtexpr.c.