pie/no-pie workaround: if linking fails, retry with -no-pie flag
On modern linux distributions the default gnu linker will expect
PIE code. Linking fails if the code is not PIE.
Passing the -no-pie option resolves this issue, but older versions
of ld don't support this flag. The workaround first tries to
link without passing -no-pie to the linker. If this fails, linking
with -no-pie is tried. If the second try fails, too, then we assume
a linking problem not caused by a missing -no-pie flag and the error
messages of the first linking attempt are printed.